Updating DynamoDB CRUD Operations.
