+++ Dockerfile
@@ -1,4 +1,5 @@
-FROM openjdk:11
+FROM openjdk:21
 EXPOSE 8080
-ADD target/Pragason-app-ecs.jar Pragason-app-ecs.jar
-ENTRYPOINT ["java", "-jar", "Pragason-app-ecs.jar"]
+ADD build/libs/aws-project-pragason-backend-api-0.0.1-SNAPSHOT.jar aws-project-pragason-backend-api-0.0.1-SNAPSHOT.jar
+
+ENTRYPOINT ["java", "-jar", "aws-project-pragason-backend-api-0.0.1-SNAPSHOT.jar"]
