+++ src/main/java/com/awsprojrct/pragason/AwsProjectPragasonBackendApiApplication.java
@@ -3,10 +3,8 @@ package com.awsprojrct.pragason;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-
 import org.springframework.context.annotation.Bean;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
 
 
 @SpringBootApplication
