+++ src/main/java/com/awsprojrct/pragason/Models/QuestionController.java
@@ -1,10 +1,10 @@
 package com.awsprojrct.pragason.Models;
 
-import com.amazonaws.services.dynamodbv2.document.Item;
 import jakarta.validation.Valid;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
 import static com.awsprojrct.pragason.constants.Constants.QuestionTBL;
@@ -28,7 +28,7 @@ public class QuestionController {
         return QuestionRepository.findAll(QuestionTBL);
     }
 
-//    Get method
+//    Get method By Id
     @GetMapping("/{id}")
     Object findById(@PathVariable String id){
 
@@ -37,17 +37,10 @@ public class QuestionController {
 
 //    POST method
     @ResponseStatus(HttpStatus.CREATED)
-    @PostMapping(" ")
+    @PostMapping(value = "/put", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
      void create (@Valid @RequestBody Question questionItem){
         questionRepository.create(questionItem);
     }
-//
-////      UPDATE method
-//    @ResponseStatus(HttpStatus.NO_CONTENT)
-//    @PutMapping("/{id}")
-//    void update(@Valid @RequestBody Question question, @PathVariable Integer id) {
-//        questionRepository.update(question, id);
-//    }
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/{id}")
