+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-server.port=8081
+server.port=8080
